Create `TreeNodeTransformer` class
In preparation for creating other decision-tree classifiers in a
nearest-neighbor context, most of the methods that were in
`RFNodeTransformer` can be moved into this new base class.  The derived
classes are mainly responsible for handling estimator-specific keywords.
